pyproject.toml
[build-system]
requires = ["poetry-core>=2"]
build-backend = "poetry.core.masonry.api"

[project]
name = "muni-data"
description = "Code to query 511 SF API to get real-time muni transit data."
version = "0.0.1"
authors = [
	{ name = "neil.bardhan", email = "indraneil.bardhan@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dynamic = ["dependencies"]

[project.scripts]
muni-data = "muni_data.main:main"

# [tool.poetry]
# package-mode = false 

[tool.poetry.dependencies]
pandas = "^2.2.0"
requests = "^2.32.2"
bump2version = "^1.0.1"
ipykernel = "^7.0.1"
jupyterlab = "^4.4.9"
python-dotenv = "^1.1.1"
google-cloud-bigquery = "^3.38.0"
google-cloud-storage = "^3.4.1"
functions-framework = "^3.9.2"
